Create trait to replace tribe_resource_url return
